APSI-58 Działające kupowanie biletów na sekcje.

+++ backend/src/main/java/apsi/team3/backend/DTOs/DTOMapper.java
@@ -176,7 +176,7 @@ public class DTOMapper {
         );
     }
 
-    public static SectionDTO toDTO(EventSection section, int ticketsBought) {
+    public static SectionDTO toDTO(EventSection section, long ticketsBought) {
         return new SectionDTO(
             section.getId(),
             section.getEvent().getId(),
